refactor(frontend): extract LibraryPanel from page.tsx (P1.f)
Moves the /library route UI and its bulk-action bar into
app/components/LibraryPanel.tsx. page.tsx drops 300+ lines.

What moved inside LibraryPanel:
- UI-local state: libraryFilter, librarySearch, selectedLibraryIds,
  isLibraryHovering, importReviewBundleInputRef.
- Derived memos: filteredLibrary, filteredLibraryWithSearch,
  showSelectionControls, allFilteredSelected.
- ~265 lines of JSX.

What stayed in HomePage:
- The actual doc-mutating handlers (handleBulkArchive, handleBulkRerun,
  handleBulkDelete, handleDeleteDocument, handleSetArchived,
  handleImportReviewBundleFiles) because they close over shared state
  like selectedDocumentId/versions/comments and need to refreshAll()
  after mutation.
- bulkProgress / isBundleImporting — owned by the parent since its
  handlers manage them; passed to LibraryPanel as read-only props.

Signature refactor on the bulk handlers:
- handleBulkArchive(archived, ids: number[])
- handleBulkRerun(targets: DocumentLibraryEntry[])
- handleBulkDelete(ids: number[])

The handlers used to close over selectedLibraryIds directly; now they
receive the selection as an argument so the state can live in the
panel. LibraryPanel clears its own selection after each bulk action
completes.

Verified: frontend 46/46 (stable across 5 full-suite reruns),
production build succeeds, /library route renders identically and
bulk/single actions still propagate to libraryEntries refresh.
